Refactor add reviewer from eua form
