+++ src/types/grbReview.ts
@@ -1,4 +1,18 @@
-import { GetGRBReviewersComparisonsQuery } from 'gql/gen/graphql';
+import {
+  GetGRBReviewersComparisonsQuery,
+  SystemIntakeGRBReviewerRole,
+  SystemIntakeGRBReviewerVotingRole
+} from 'gql/gen/graphql';
+
+export type GRBReviewerFormFields = {
+  userAccount: {
+    username: string;
+    commonName: string;
+    email: string;
+  };
+  votingRole: SystemIntakeGRBReviewerVotingRole;
+  grbRole: SystemIntakeGRBReviewerRole;
+};
 
 export type GRBReviewerComparisonIntake =
   GetGRBReviewersComparisonsQuery['compareGRBReviewersByIntakeID'];
